fix type annotationns for parameters

+++ 0x01-python_async_function/3-tasks.py
@@ -6,7 +6,7 @@ import asyncio
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-def task_wait_random(max_delay: float) -> asyncio.Task:
+def task_wait_random(max_delay: int) -> asyncio.Task:
     """
     takes an integer max_delay and returns a asyncio.Task.
     """
